sql/sql/sem/cast: extract a package from tree
This cast logic is both complex from a domain perspective and pretty
self-contained. It probably could have been left in tree, but no reason to do
that.

Release note: None
