+++ pkg/sql/catalog/tabledesc/fix_cast_for_style_visitor.go
@@ -16,6 +16,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/descpb"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/schemaexpr"
+	"github.com/cockroachdb/cockroach/pkg/sql/sem/cast"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/volatility"
 	"github.com/cockroachdb/cockroach/pkg/sql/sessiondata"
@@ -101,7 +102,7 @@ func (v *fixCastForStyleVisitor) VisitPost(expr tree.Expr) tree.Expr {
 		innerExpr := expr.Expr.(tree.TypedExpr)
 		outerTyp := expr.ResolvedType()
 		innerTyp := innerExpr.ResolvedType()
-		vol, ok := tree.LookupCastVolatility(innerTyp, outerTyp, &sd)
+		vol, ok := cast.LookupCastVolatility(innerTyp, outerTyp, &sd)
 		if !ok {
 			v.err = errors.AssertionFailedf("Not a valid cast %s -> %s", innerTyp.SQLString(), outerTyp.SQLString())
 		}
