+++ pkg/sql/colexec/colexecbase/BUILD.bazel
@@ -69,6 +69,7 @@ go_test(
         "//pkg/sql/execinfrapb",
         "//pkg/sql/faketreeeval",
         "//pkg/sql/randgen",
+        "//pkg/sql/sem/cast",
         "//pkg/sql/sem/tree",
         "//pkg/sql/types",
         "//pkg/testutils/skip",
