+++ pkg/sql/opt/optbuilder/union.go
@@ -14,6 +14,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/opt/memo"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgcode"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgerror"
+	"github.com/cockroachdb/cockroach/pkg/sql/sem/cast"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/types"
 )
@@ -159,7 +160,7 @@ func determineUnionType(left, right *types.T, clauseTag string) *types.T {
 			// If the right type is "larger", use it.
 			src, tgt = left, right
 		}
-		if !tree.ValidCast(src, tgt, tree.CastContextExplicit) {
+		if !cast.ValidCast(src, tgt, cast.CastContextExplicit) {
 			// Error if no cast exists from src to tgt.
 			// TODO(#75103): For legacy reasons, we check for a valid cast in
 			// the most permissive context, CastContextExplicit. To be
