+++ pkg/sql/pg_catalog.go
@@ -36,6 +36,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgerror"
 	"github.com/cockroachdb/cockroach/pkg/sql/privilege"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/builtins"
+	"github.com/cockroachdb/cockroach/pkg/sql/sem/cast"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree/treecmp"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/volatility"
@@ -481,8 +482,8 @@ https://www.postgresql.org/docs/9.6/catalog-pg-cast.html`,
 	schema: vtable.PGCatalogCast,
 	populate: func(ctx context.Context, p *planner, _ catalog.DatabaseDescriptor, addRow func(...tree.Datum) error) error {
 		h := makeOidHasher()
-		tree.ForEachCast(func(src, tgt oid.Oid, cCtx tree.CastContext, ctxOrigin tree.ContextOrigin, _ volatility.V) {
-			if ctxOrigin == tree.ContextOriginPgCast {
+		cast.ForEachCast(func(src, tgt oid.Oid, cCtx cast.CastContext, ctxOrigin cast.ContextOrigin, _ volatility.V) {
+			if ctxOrigin == cast.ContextOriginPgCast {
 				castCtx := cCtx.PGString()
 
 				_ = addRow(
