+++ pkg/sql/randgen/BUILD.bazel
@@ -28,6 +28,7 @@ go_library(
         "//pkg/sql/rowenc",
         "//pkg/sql/rowenc/keyside",
         "//pkg/sql/rowenc/valueside",
+        "//pkg/sql/sem/cast",
         "//pkg/sql/sem/tree",
         "//pkg/sql/sem/tree/treebin",
         "//pkg/sql/sem/tree/treecmp",
