+++ pkg/sql/sem/cast/cast_map_gen.sh
