+++ pkg/sql/sem/tree/BUILD.bazel
@@ -138,12 +138,12 @@ go_library(
         "//pkg/sql/catalog/catconstants",
         "//pkg/sql/lex",
         "//pkg/sql/lexbase",
-        "//pkg/sql/oidext",
         "//pkg/sql/pgwire/pgcode",
         "//pkg/sql/pgwire/pgerror",
         "//pkg/sql/pgwire/pgnotice",
         "//pkg/sql/privilege",
         "//pkg/sql/roleoption",
+        "//pkg/sql/sem/cast",
         "//pkg/sql/sem/catid",
         "//pkg/sql/sem/tree/treebin",
         "//pkg/sql/sem/tree/treecmp",
@@ -192,7 +192,6 @@ go_test(
     srcs = [
         "as_of_test.go",
         "cast_map_test.go",
-        "cast_test.go",
         "col_types_test.go",
         "collatedstring_test.go",
         "compare_test.go",
@@ -240,7 +239,6 @@ go_test(
         "//pkg/settings/cluster",
         "//pkg/sql/catalog/typedesc",
         "//pkg/sql/faketreeeval",
-        "//pkg/sql/oidext",
         "//pkg/sql/opt/exec/execbuilder",
         "//pkg/sql/opt/optbuilder",
         "//pkg/sql/opt/xform",
@@ -249,6 +247,7 @@ go_test(
         "//pkg/sql/pgwire/pgerror",
         "//pkg/sql/randgen",
         "//pkg/sql/sem/builtins",
+        "//pkg/sql/sem/cast",
         "//pkg/sql/sem/tree/treebin",
         "//pkg/sql/sem/tree/treecmp",
         "//pkg/sql/sem/tree/treewindow",
@@ -272,7 +271,6 @@ go_test(
         "//pkg/util/timeutil/pgdate",
         "@com_github_cockroachdb_apd_v3//:apd",
         "@com_github_cockroachdb_datadriven//:datadriven",
-        "@com_github_cockroachdb_errors//:errors",
         "@com_github_lib_pq//oid",
         "@com_github_stretchr_testify//assert",
         "@com_github_stretchr_testify//require",
