+++ pkg/sql/sem/tree/cast_map_test.go
@@ -18,6 +18,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgcode"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgerror"
 	"github.com/cockroachdb/cockroach/pkg/sql/randgen"
+	"github.com/cockroachdb/cockroach/pkg/sql/sem/cast"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/volatility"
 	"github.com/cockroachdb/cockroach/pkg/sql/types"
@@ -37,7 +38,7 @@ func TestCastMap(t *testing.T) {
 	rng, _ := randutil.NewTestRand()
 	evalCtx.Planner = &faketreeeval.DummyEvalPlanner{}
 
-	tree.ForEachCast(func(src, tgt oid.Oid, _ tree.CastContext, _ tree.ContextOrigin, _ volatility.V) {
+	cast.ForEachCast(func(src, tgt oid.Oid, _ cast.CastContext, _ cast.ContextOrigin, _ volatility.V) {
 		srcType := types.OidToType[src]
 		tgtType := types.OidToType[tgt]
 		srcDatum := randgen.RandDatum(rng, srcType, false /* nullOk */)
