+++ pkg/sql/sem/tree/type_check.go
@@ -18,6 +18,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/server/telemetry"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgcode"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgerror"
+	"github.com/cockroachdb/cockroach/pkg/sql/sem/cast"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree/treecmp"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/volatility"
 	"github.com/cockroachdb/cockroach/pkg/sql/sessiondata"
@@ -494,15 +495,15 @@ func resolveCast(
 		return nil
 
 	default:
-		cast, ok := LookupCast(castFrom, castTo, intervalStyleEnabled, dateStyleEnabled)
+		c, ok := cast.LookupCast(castFrom, castTo, intervalStyleEnabled, dateStyleEnabled)
 		if !ok {
 			return invalidCastError(castFrom, castTo)
 		}
-		if !allowStable && cast.Volatility >= volatility.Stable {
+		if !allowStable && c.Volatility >= volatility.Stable {
 			err := NewContextDependentOpsNotAllowedError(context)
 			err = pgerror.Wrapf(err, pgcode.InvalidParameterValue, "%s::%s", castFrom, castTo)
-			if cast.VolatilityHint != "" {
-				err = errors.WithHint(err, cast.VolatilityHint)
+			if c.VolatilityHint != "" {
+				err = errors.WithHint(err, c.VolatilityHint)
 			}
 			return err
 		}
